postgres/requirements.txt
ipykernel==6.23.2
fastapi==0.85.0
uvicorn==0.18.3
psycopg2_binary==2.9.4
python-dotenv==0.21.0
pipreqs==0.4.13
pydantic==1.10.9
email-validator==1.3.0
passlib==1.7.4
bcrypt==4.0.0
python-multipart==0.0.5